// src/@core/layouts/types.ts
import { ReactNode } from 'react';
import { Settings } from 'src/@core/context/settingsContext';

export type ContentWidth = 'full' | 'boxed';

export type ThemeColor = 'primary' | 'secondary' | 'error' | 'warning' | 'info' | 'success';

export type NavLink = {
	path?: string;
	title?: string;
	action?: string;
	subject?: string;
	disabled?: boolean;
	badgeContent?: string;
	externalLink?: boolean;
	openInNewTab?: boolean;
	icon?: string | string[] | ReactNode;
	badgeColor?: 'default' | 'primary' | 'secondary' | 'success' | 'error' | 'warning' | 'info';
	children?: any;
	isNav?: boolean;
	sectionTitle?: string;
	key?: string;
	dropdownItems?: [
		{
			title: string;
			path: string;
			key: string;
		}
	] | any
};

export type VerticalNavItemsType = NavLink[];
export type VerticalNavItemType = NavLink;

export type LayoutProps = {
	hidden: boolean;
	settings: Settings;
	children: ReactNode;
	verticalNavItems?: VerticalNavItemsType;
	scrollToTop?: (props?: any) => ReactNode;
	saveSettings: (values: Settings) => void;
	footerContent?: (props?: any) => ReactNode;
	verticalAppBarContent?: (props?: any) => ReactNode;
	verticalNavMenuContent?: (props?: any) => ReactNode;
	verticalNavMenuBranding?: (props?: any) => ReactNode;
	afterVerticalNavMenuContent?: (props?: any) => ReactNode;
	beforeVerticalNavMenuContent?: (props?: any) => ReactNode;
};

export type BlankLayoutProps = {
	children: ReactNode;
};